Handles feedback from wordpress plugin submission progress:
'Adblockers to Bitcoin' would be okay.

So would 'owocki's Bitcoin Adblock' or 'Block Ads to Bitcoin'

That said, I'd use 'owocki bitcoin adblock' as that gives you the URL slug 'owocki-bitcoin-adblock' :)

With this, yeah, Adblock puts you in a bad place, but since you're making a plugin that DOES adblocking, you would run into trademark laws. We kind of want to avoid that for you.

You can use adblock and bitcoin in the name, they just can't be the FIRST term in your name.

Please use more generic defaults in your plugin. When you default to existing media or content, it often inadvertently causes people to have their sites displaying content they did not wish to show. You can have your code show http://example.com or your shortcode not display information, instead of forcing the display of content.

$address = '1A3KJ7f3YAS9jvhPjotLEzEMd1dNKZ9EPE';

Don't default to YOUR address please :)

All plugins should have unique function names, defines, and classnames. This will prevent your plugin from conflicting with other plugins or themes.

For example, if your plugin is called "Easy Custom Post Types", then you might prefix your functions with ecpt_{your function name here}. Similarly a define of LICENSE would be better done as ECPT_LICENSE. You can use namespaces instead, however make sure that those also are unique. A namespace or class of 'MyPlugin' is NOT actually all that unique, you see.

Please update your plugin to use more unique names.

Functions like get_default_copy are not unique.

Your plugin is using `<style>` and/or `<link>` tags to insert CSS/JS

```
<script type="text/javascript">document.always_show_bitcoin_solicitations=1;</script>':'').'
    <script src="//netdna.bootstrapcdn.com/bootstrap/3.0.3/js/bootstrap.min.js"></script>
    <script src="'.get_plugin_staticfiles_url().'bitcoinaddress.js/dist/demo.js"></script>
    <script src="'.get_plugin_staticfiles_url().'FuckAdBlock/fuckadblock.js"></script>
    <script src="'.get_plugin_staticfiles_url().'adblock-to-bitcoin.js"></script>
```

You should be using the built in functions for this:

https://codex.wordpress.org/Function_Reference/wp_enqueue_script
https://codex.wordpress.org/Function_Reference/wp_enqueue_style

If you're trying to enqueue on the admin pages you'll want to use the admin enqueues

https://codex.wordpress.org/Plugin_API/Action_Reference/admin_enqueue_scripts
https://codex.wordpress.org/Plugin_API/Action_Reference/admin_print_scripts
https://codex.wordpress.org/Plugin_API/Action_Reference/admin_print_styles

Examples:

//netdna.bootstrapcdn.com/bootstrap/3.0.3/js/bootstrap.min.js

Offloading images, js, css, cgi, and other scripts to Google (or jquery.com or anywhere else frankly) is disallowed because you're introducing an unnecessary dependency on another site.  If the file you're trying to use isn't a part of WordPress Core, then you should include it -locally- in your plugin, not remotely. If the file IS included in WordPress core, please call that instead.

The one exception to this rule is if your plugin is performing a service. We will permit this on a case by case basis, however since this can be confusing, we have some examples of what are not permitted:

* Offloading jquery CSS files to Google - You should include the CSS in your plugin.
* Inserting an iframe with a help doc - A link, or including the docs in your plugin is preferred.
* Calling images from your own domain - They should be included in your plugin.

Here are some examples of what we would permit:

* Calling font families from Google or their approved CDN (if GPL compatible)
* API calls back to your server to process possible spam comments (like Akismet)
* Offloading comments to your own servers (like Disqus)
* oEmbed calls to a service provider (like Twitter or YouTube)

Please remove this dependency from your plugin and, if possible, include all files within the plugin (that is not called remotely). If instead you feel you ARE providing a service, please re-write your readme.txt in a manner that explains the service, the servers being called, and if any account is needed to connect.

NOTE! It's OKAY to call FONTS. That means this is allowed: //netdna.bootstrapcdn.com/font-awesome/4.0.3/css/font-awesome.css

Just not the other stuff :)

We do not permit plugins to insert Powered By attribution in plugins on any front-facing aspect of the site, without the user's express permission. This includes credits, be they links or not, powered by attribution, css-hidden links, or any other advertising whatsoever.

You are permitted to have an HTML comment, or to have the attribution be opt-in, however we're serious about Opt IN. This means the user must check a box to indicate that yes, they want to show the credit on their site. Under no circumstances are you permitted to display the attribution without permission.

The sole exception to this rule is service ware (such as Twitter, Disqus, or YouTube), which includes the attribution on their own site, not in the plugin.

Please add a nonce to your POST calls to prevent unauthorized access.

https://codex.wordpress.org/WordPress_Nonces

The way your plugin is referencing other files is not going to work with all setups of WordPress.

Example:

return get_site_url().'/wp-content/plugins/adblock-to-bitcoin/static/';

When you hardcode in paths, or assume that everyone has WordPress in the root of their domain, you cause anyone using 'Giving WordPress it's own directory' (a VERY common setup) to break. In addition, WordPress allows users to change the name of wp-content, so you would break anyone who choses to do so.

Please review http://codex.wordpress.org/Determining_Plugin_and_Content_Directories and update your plugin accordingly. And don't worry about supporting WordPress 2.x or lower. We don't encourage it nor expect you to do so, so save yourself some time and energy.

----

Please make sure you've addressed ALL issues brought up in this email. When you've corrected your code, reply to this email with the updated code attached as a zip, or provide a link to the new code for us to review. If you have questions, concerns, or need clarification, please reply to this email and just ask us.